Funnel Retail
Using BigQuery and Google Analytics 4, we conducted an analysis of advertising campaigns on Facebook and Google Ads for the year 2021. The data was prepared for reporting in BI systems, conversions from session start to purchase were calculated, and the effectiveness of various landing pages was investigated. Additionally, we explored the correlation between user engagement and purchases.